Raise useful exception when prison not found
In issues #208 and #209, we see exceptions raised when there is no
prison data for the prison name supplied. It's more helpful to raise an
exception earlier that tells us that the key does not exist in the
prison data hash.

Instead of:

    ActionView::Template::Error: undefined method `[]' for nil:NilClass

the exception raised by an unknown prison name is now:

    KeyError: key not found: "bar"

This should assist us in tracking down the root cause of these errors.