Update specs to current RSpec layout
There is now both a spec_helper and a rails_helper: the former is useful
when you want to test something without the massive overhead of loading
all of Rails first. For all existing specs, they have been changed to
rails_helper so that they work, though this may not be necessary in all
cases.

The default configuration for RSpec 3 does not add a 'describe' method
to Object. This is probably a good thing, but means that specs must now
start with 'RSpec.describe', rather than just 'describe'.