Add a helper to render markdown
This uses the same options as the Redcarpet template implementation,
with the addition of the filter_html option to disallow arbitary HTML
input. This doesn't matter at the moment, as we intend to use it to
render templates, but is set in case it is used in other ways in the
future.