Make spec pass when month starts on a Monday
Says the comment:

> If the week starts on Monday, don't run this test - "yesterday" will
> be beyond the first row of the calendar.

Weeks are fairly constant, of course: what they meant was that the month
starts on a Monday. This is an unreasonable restriction, so we'll
instead attempt to click the last date in the calendar, as this will
always be more than 14 or 28 days (for instant or deferred).