Ensure datepicker specs run after 2015/06/09
`datepicker_spec` `Choosing a date in the past` failed for two reasons:

- The `Timecop.travel` call was *after* the page was rendered, so had
  no impact on the rendering.
- Travelling to the past resulted in cookies that expired in the past
  being issued to the test browser. Because the test browser doesn't
  respect our time-travelling, it considered these cookies expired
  instantly.

To avoid the session expiring, we should always travel to a suitable
date in the future (a Tuesday, so we can test booking a prior working
day in `Choosing a date in the past`).

Also ensure we pass the server's idea of “today” to the JavaScript (via
the views' data-today attributes) to avoid weird miscalculations of
booking window start- and end-dates on the client-side.