Clarify intent of schedule tests
We shouldn't have to do date maths in our head to understand the tests,
so remove the magic numbers.

It also turns out that the code doesn't behave at all as it should, so
this change is mostly to clarify what the code *actually* does. That
way, when we come to fix the behaviour in future, there's a good example
of how to communicate and test that change.

Also clarify some parameters and make date construction consistent.