Extends the `PrisonDay` class with a `.visiting_day?` method
Provides a public `.visiting_day?` method that allows a given
day to be queried as to whether or not the prison provided
(when the class was constructed) allows visitation or not.

Accompanied by a spec which provides tests for a number of
different scenarios(public holidays, specific
unbookable days, anomalous dates and days not in the prisons
expected slots).