Draft `PrisonSchedule` class
The intention here is to replace the `Schedule` class found
under lib.

Exposes two public methods, one for gathering a date for when
the confirmation is expected to be delivered and the other to
provide a list of acceptable dates visits may occur on.

Q: Are the ranges present in the class sufficient?

Specs to be added once range question has been answered.