Add a `prison_name` helper method
This was accidentally removed in SHA:73ea26f3a81d719603756c746b27085697b6b78c

-  def prison_name(source=visit)
-    source.prisoner.prison_name
-  end

Adding it back as it's required in the prisoner_details template
for rendering out error messages related to the state of the prison.

Removed the auto-seeding argument with the current visit as it's
unnecessary.

Commit includes a test in the visit_helper_spec. If a test had
been around when the code was first committed it probably wouldn't
have been shown the door. :sad face: