Confirmation email date will ignore today as a staff working day
The lead days specified by the prison must exclude the current day
if it is classified by the prison as a working day. The lead days
should be calculated as N days from tomorrow & where those days
cannot be the expected date either.

For example, if a prison has a lead days value of 3 and we are booking
on a Monday then we would expect a confirmation email date on the
Friday.

Mon - [Tue] - [Wed] - [Thu] - Fri
 X      1       2       3     ^^^