Correct the order for 'bottom' leaderboard rankings
This ensures that the list returned from the `leaderboard`
endpoint in the `GeckoboardController` is presented in the 
correct order.

The head of the list will now yield the worst performer with
each consecutive element improving.

Prior to this fix the list was returned in descending
performance order(the worst performer would be the 10th element).