Fix calculation of confirmation date
Previously, the system calculated the confirmation date as the working
day after the lead time. It should be the last day of the lead time.

For example, if a prison processes visits Monday to Friday, and Saturday
is a visiting day, then it should be possible to book a visit for the
coming Saturday:

    Tuesday   | Make booking
    Wednesday | Processing
    Thursday  | Processing
    Friday    | Processing & Confirmation
    Saturday  | Visit

Previously, because of the extra day, the confirmation would not be sent
until Monday, and visits would not be possible until Tuesday.