Specify behaviour when Sendgrid is not configured
When the SMTP user name and password are missing, as is the case in
development by default, the Sendgrid helper should not contact Sendgrid,
but should instead say that the email address is OK (i.e. it has passed
the bounced? and spam_reported? checks). This makes local development
much easier.