Allow in-memory objects to be (de)queued
This adds a custom to_global_id method to the NonPersistedModel module
defined in the previous commit, and a corresponding locator that is able
to rehydrate these URIs.

This is implemented by using a URL-safe base-64-encoded version of a
JSON representation of a hash of the object's attributes, the latter
being generated by Virtus. This encoded string is used in place of the
ID in the normal Global ID URI scheme (gid://[locator]/[class]/[id]).

The locator (registered as 'pvb') then instantiates the specified class
using the attributes from the ID portion of the URI.

Note: the behaviour of nested object attributes is not specified.

See https://github.com/rails/globalid for more details on how Global IDs
work.