Add Zendesk to healthcheck.json
The Zendesk API library raises an exception if the API is unavailable
when attempting to save a ticket, but does not raise an exception for
read actions. However, after experimentation, we have determined that
the tickets collection proxy returns a count of -1 if it could not
connect to the API. Thus, we check whether there are zero or more
tickets to determine that the API is available.