Do not send emails to prisons when smoke testing
We can determine if the user is a 'fake' smoke test user
by the email address being used by the visitor. If we
determine that a smoke test is being performed we change
the emails recipient so that it gets sent to our smoke test
email account and we can later perform assertions on it.