Add queues to healthcheck & remove SMTP & Zendesk
Now that we are going to send emails via queues, we don't actually care
about transient unavailability of SMTP servers or the Zendesk API.
Instead, we care how many items are on the queues, and how old they are.

The healthcheck response now looks like this:

    {
      "checks": {
        "database": true,
        "mailers": false,
        "zendesk": false
      },
      "queues": {
        "mailers": {
          "oldest": "2015-08-24T11:19:21.993Z",
          "count": 2
        },
        "zendesk": {
          "oldest": "2015-08-24T11:19:21.995Z",
          "count": 1
        }
      }
    }

The mailers and zendesk checks pass if there are no stale items in those
queues, which is computed by checking whether the oldest queue item is
more than ten minutes old.