Add failing tests for skipping a step
If a visitor jumps in at a later page, either by using the back button,
or by otherwise saving the URL, they can arrive at an invalid state in
which no prison is associated with the visit. This should take them
gently back to the start: instead, they get a 500 error.