Check integrity of session visit data at each step
After the first step, the session must contain a visit, which has a
prisoner, who has a prison name. Without this, the prison data cannot be
retrieved and an exception results.

Now, if the session doesn't contain this information, we'll redirect the
visitor back to the first step with a message and get them to start
again.

This won't solve every possible 500 error, but it should make a
difference.