Make email validation independent of model
Rails-style validation isn't enough for what we want to do, which is to
be able to tell whether certain aspects of invalidity can be overridden.
Instead of shoehorning this in, we'll make the EmailValidator a
separate object that can be interrogated by the Visitor model. In
addition, we need to change the Feedback model, which also uses email
validation.