Set start of process to be `/prisoner` page
Development and staging environements redirect any hits to the home page to
the live environment! Any subsequent steps would therefore be run
on that environment and might lead to hmmm.. real life visit
requests. So long as the envirment variable with the host is set
correctly its quite acceptable to just skip this step and hit the first
part of the form.

Concerns: If for some reason a page was to be injected before
the prisoner details page it could be problematic - for the sake of this
current smoke test implementation we are not too worried about this.