Remove unnecessary foreign data source check
It seems unnecessary to make a call out to get the bank
holiday data every time the suite of specs is run. Besides
this the actual test makes no assertion! & the attempt at
an assertion incorrectly uses sets `proper_superset?`
method which is inappropriate for comparing sets that
contain exactly the same items.

Imagine the following:

   Set A = [ 1, 2, 3]
   Set B = [ 1, 2, 3, 4]

In the example above Set B is a 'proper superset' of
Set A as it contains at least one more element that is not
in Set A.

Moving on..

Whenever the data is updated via the rake task we would
spot inconsistencies in the data when the specs run.

A test has been added to assert the data contains the
expected keys.