Ensure Content-Transfer-Type is 'quoted-printable'
Sendgrid enforces line length limits of 998 characters. This
is according to the RFC spec: http://tools.ietf.org/html/rfc5322
see section 2.1.1 for more info.

We have noticed that ActionMailer has some inconsistencies
with what Content-Transfer-Type header is set for Prisoner
& Visitor emails. The VisitorMailer emails are nearly always
set as 'quoted-printable' whereas the Prison emails are '7bit'.
Sending through Gmail & GSI does not alter the contents of the
email whereas Sendgrid will enforce the rulings of the RFC
spec unless the transfer encoding is set to 'quoted-printable' or
'base64' and add line breaks which will break long links as found
in the PrisonMailer emails.

This addition adds a filter than can be mixed into any of the
mailers we have in the application and will ensure the
Content-Transfer-Type header is always set to 'quoted printable'

Quoted printable is an encoding that allows us to transmit 8 bit
data over a 7 bit medium and limits line length to 76 characters
by using a '=' as a separator. It produces human readable output
that is rendered correctly by the likes of Gmail for both plain
text and html emails.

This issue was brought to our attention as the smoke tests found
in PR #268 override the smtp settings with the default action
mailer settings which use Sendgrid - the emails had broken links.

This is a shameful hack.