Add removal of email's from spam & bounce lists
Makes use of Sendgrid Toolkits deletion feature which underneath uses
Sengrids WEB API endpoints.

Both of these methods are required in order for a user to override the
default spam and bounce checks if they believe ther email address ahs
been flagged in the past.

The toolkit raises an exception when no email address is found in the
spam or bounce list - we dont want to treat this as exceptional
circumstance as it will cause the delivery mechanism(background job i.e
Sidekiq) to fail and retry. Any other errors which might occur such as
API Errors or parsing errors we expect to escalate to the job level thus
failing the job an allowing it to be retried.