Ensure visitors receipt email attempts Sendgrid resets
We utilise the Sendgrid API to attempt to remove the users email
address from spam or bounce lists before sending the email so
long as the option has been selected.

The email validator allows us to query whether or not we can attempt
to remove the visitors email address from the spam report list or
the bounces list.