Cache users choice for performing bounce or spam reset
Adds a hidden field to the visitor form so we can persist the
override type that has been displayed to the user via the error
message and not have to ask Sendgrid again later on.

Extracted a `Visitor` decorator that handles determining the
reset(spam or bounce) & then performing it.