Remove documentary comments/associated whitespace
I started to do this as part of the rubocop LineLength TODO. On
reflection, it seemed cleaner to do it on a separate branch as not all
of them were excessively long.  Nevertheless, we don't want them in the
codebase.  Where removal would result in an empty file, as with
config/initializers/backtrace_silencers.rb, I have left the opening
comment (almost always a warning to restart the server if the file
changes).