Remove browserstack rake task and integration
Steve thinks this was the original attempt at developing smoke tests. It
was commented out in .circle.yml with an instruction to uncomment it
when it was made to work.  I could not get it to run at all locally and
the jenkins job it was supposedly powering had not succeeded once in
over 300 runs since early November 2014.

I am now going to add a ticket to the backlog to re-implement it using a
more ruby-like approach (i.e. not spawning external java classes to
proxy communication to Browserstack by overriding the port assigments of
a webbrick server, &tc. &tc.)