Add specs for entry points with serialised data
This application does not use a database (almost; but see the junk
drawer metrics database). Instead, it passes serialised and encrypted
data around in URLs in emails.

This is very painful.

It means that the application is very fragile in the face of any change
to classes, namespaces, etc., because the current state of the
application will still be exposed to serialised objects from previous
versions.

In order to give us more confidence, this feature ensures that some
pickled data can be loaded. The fixtures were created using a temporary
scenario, which is listed below for posterity.

    scenario 'regenerate' do
      visit edit_prisoner_details_path

      complete_prisoner_details 'Cardiff'
      click_button 'Continue'

      complete_visitor_details
      click_button 'Continue'

      complete_visit_details
      click_button 'Continue'

      click_button 'Send request'

      urls = ActionMailer::Base.deliveries.flat_map { |d|
        d.text_part.body.decoded.lines.grep(/http/)
      }.uniq

      p VisitMetricsEntry.last
      puts *urls

      visit urls.first.gsub(%r{http://localhost:3000}, '')

      ActionMailer::Base.deliveries.clear
      choose 'confirmation_outcome_slot_0'
      fill_in 'confirmation_vo_number', with: '12345678'
      click_button 'Send email'
      save_and_open_page

      urls = ActionMailer::Base.deliveries.flat_map { |d|
        d.text_part.body.decoded.lines.grep(/http/)
      }.uniq

      p VisitMetricsEntry.last
      puts *urls
    end