Stop passing around marshalled Ruby objects
The use of marshalled Ruby objects is problematic: as they exist in
links in emails outside our control, our application must be able to
deal with objects from the past at any time. However, these are not
normally visible to the test suite nor when testing the application
manually.

We have had to fight this problem once, and we do not wish to repeat it.
Thus, we change the serialisation and encryption to use a JSON
representation instead. Future changes should hopefully require only
some munging of the structure of the hash at worst.

As noted above, previous objects still exist, so we can't forget about
them entirely. YET. We treat them as a special case: if the token, once
decrypted, can't be interpreted as JSON, we'll use Marshal.load as
before.