Extract haml_io << messiness
Using ERB templates (the .html is misleading here) is at least less
awful than lines upon lines of `haml_io << "var …`. I'm sure that
further improvement is possible, but for now I just want to break down
the templates a little.

There's an obvious wart in the fact that we're using `.html_safe`, but
this is exactly what was happening before; without it, quotes are
escaped and the output isn't valid JSON at all.