Simplify error-highlighting element wrappers
A common activity in the templates is to wrap a set of elements in a
div, give it a class of 'field' or 'group', and add the
'validation-error' class if the field has an error.

There are a couple of oddities remaining due to the dual
visitor_not_listed/unlisted and visitor_banned/banned fields.