Extract date-of-birth to partials
These are now ERB templates so that we can use line breaks to organise
the excessively-long code in a more readable manner.

There is a lot of duplication between the prisoner and visitor
templates. It should be possible to merge these, but it's not
immediately clear how to do this without an unreasonable amount of
configuration.

The minimum and maximum values should be obtained from the relevant
object.