Get DOB input min/max from object
The arbitrary `Date.today.year - 15` was not set based on any logic:

"In England & Wales the age of criminal responsibility is set at 10.
Young offenders aged 10 to 17 (i.e. up to their eighteenth birthday) are
classed as a juvenile offender. Between the ages of 18 and 21 (i.e. up
to their twenty-first birthday) they are classed as young offenders.
Offenders aged 21 and over are known as adult offenders."

So let's just use the same maximum everywhere, which always works.

We can also simplify the field indices instead of calculating
`count - i` all over the place.