Make visitor_mailer_spec shorter and use Prison
I broke out the =~ 100 lines of let definitions and hooks into a shared
concern.  The spec, proper is still over 300 lines long (!!!), but this
helped clear it up a bit.  It really needs to be broken down into a
number of much shorter, clearer specs, but that was outside the scope of
this excercise.  I changed it to use the Prison model instead of the old
prison_data hash.