Make the intent of the spec clearer
It was failing in this iteration owing to the fact that the passing and
failing tests were grouped in the same `it` block. It is unclear why
they were passing on master as they should not have been.  Overall, the
intent behind the spec was obscured by the wording of the blocks and
this made bottoming the problem more difficult. I broke the passing and
failing versions out into their own specs to ensure that the data set
under test was not being handed off between them.