Ensure key validation executes in constant-time
To avoid nefarious parties carrying out timing attacks to work out the
length of our access keys, we should ensure that the checks always run
in constant(ish)-time. Rather than developing a constant-time
string-padding algorithm (as I was trying to do), we decided we should
just pad the execution time to a fixed time.