Move ParamUtils to a better-named module
It's my contention that nothing should ever be called Utils. In this
case, it has one clear purpose: to provide a function that trims
whitespace from parameters.

It's not clear *why* we need to trim whitespace from parameters,
however, or why it's only used in a couple of places rather than
everywhere.