Refactor prison sources into individual files
I was getting a lot of merge errors on the monolithic `prison_data.yml`
file.  These often caused subtle and difficult-to-resolve errors.  For
example, it merged out the name and nomis id of Crookham Wood.  This
caused tests using Cardiff, the prison just before to fail owning the
the fact that the Crookham Wood attributes overrode the Cardiff
attributes silently.  After discussion, we decided that a
prison-per-file data strucutre would be easier to maintain while not
adding much overhead (as the data is only loaded once).