Remove legacy data fixes
These prison names were last changed in 2014, and we've checked the
metrics and verified that these fixes are no longer used.

Whilst we may need to handle prison name changes in the future, we will
do this in a better way than by dumping a hash of strings directly into
the controller.