Remove date input manipulation
Virtus will automatically handle date inputs of the form

    { day: '01', month: '12', year: '2000' }

so we don't need to replicate the complicated 1i/2i/3i Rails format.

There is one small wrinkle, however: if empty strings are assigned,
Virtus will return a hash of the inputs instead of a date object. We
therefore override the date_of_birth method to return either a Date or
nil.