Refactor Visit Date Format
We found that  this app has many different date
formats such as "25 October 2015",
"Sunday 25 October 2015", "30/10/2015",
and both 12 and 24-hour times. Time periods are
shown as "1:30pm to 3:00pm", "1:30pm, 1 hr 30 mins",
and "13:45-16:00".

According to GDS design pattern we will use the format
"Sunday 25 October" for Visit and "25 October 2015" for
the Date of Birth. For time slots we will use
"1:30pm for 1h".