Add estate to every prison
The nomis_id is not unique amoung the different organisational units
that comprise some prisons (Winchester Remand Only and Winchester
Convicted Only, for example). However, it is used for calculating
metrics. This causes errors as it unclear which unit the metric is
related to.  Adding an organisational unit of 'estate' should resolve
this as it will allow the report to show the higher-level organisation
the nomis_id actually appies to ('Winchester' here, for instance).