Move integrity check and log missing info
We are experiencing 500 errors on the live service that *appear* to be
happening when existing visits are deserialised. Visits lose their
prisoner_name and prison.slots attributes.  This commit removes the
bulk of the Visit integrity checking logic to a service and uses
the `ensure_visit_integrity` method to call a series of logging methods
that look for missing prisoner and prison associations.  Hopefully, this
will give us some insight on where these errors are coming from.