Put metrics DB defaults into consistent state
Because some migrations were edited after the fact (in dca3b2e and
7a4d9a6), the database on the CI server was in a state that didn't
reflect the schema. Previous attempts to fix this updated the data, but
not the schema.

This migration sets any missing values, ensures that the defaults are
correct, and adds NOT NULL constraints to the relevant columns so that
they stay correct.

The lesson here is never to modify a migration that has been released
into the wild.

We delete the two previous commit that attempted to fix this because it
makes no useful change to the state, and is superseded by this.