Fix unbookable and anomalous conflicts on LCI
All of the dates had passed, but I am not confident that the unbookable
and slot_anomalies data structures are not being accessed directly.
The VisitHelper is delegating slot_anomalies to Prison.  It *should* be
using the anomalous_dates method, instead.

I decided against addressing this now as the potential for subtle
breakage feels high and the workaround of including blank data
structures in the yaml solved the problem.

It is worth noting that the specs fail if the unbookable and
slot_anomalies values are left blank.